some fixes around the project

+++ src/main/java/com/onlinemarket/core/service/UserService.java
@@ -1,8 +1,12 @@
 package com.onlinemarket.core.service;
 
+import com.onlinemarket.core.exceptions.ApiExceptions.ConflictException;
 import com.onlinemarket.core.exceptions.auth.UserAlreadyExistsException;
+import com.onlinemarket.core.exceptions.general.BadRequestException;
 import com.onlinemarket.core.exceptions.repository.ResourceNotFoundException;
+import com.onlinemarket.core.model.Order;
 import com.onlinemarket.core.model.User;
+import com.onlinemarket.core.repo.OrderRepo;
 import com.onlinemarket.core.repo.UserRepo;
 import com.onlinemarket.rest.dto.user.UserDTO;
 import com.onlinemarket.rest.dto.user.UserRequestDTO;
@@ -15,15 +19,18 @@ import static java.util.stream.Collectors.toList;
 
 @Service
 public class UserService {
-    private UserRepo userRepo;
+    private final UserRepo userRepo;
 
-    public UserService(UserRepo userRepo) {
+    private final OrderRepo orderRepo;
+
+    public UserService(UserRepo userRepo, OrderRepo orderRepo) {
         this.userRepo = userRepo;
+        this.orderRepo = orderRepo;
     }
 
     public List<UserDTO> findAll() {
         List<User> users = userRepo.findAll();
-        return users.stream().map(user -> new UserDTO(user)).collect(toList());
+        return users.stream().map(UserDTO::new).collect(toList());
     }
 
     public UserDTO findById(String id) {
@@ -47,6 +54,11 @@ public class UserService {
 
     public UserDTO updateUser (String id, UserRequestDTO payload) {
         Optional<User> user = userRepo.findUserById(id);
+
+        if(!(payload instanceof UserRequestDTO)){
+            throw new BadRequestException("Request body not matching the expected type");
+        }
+
         if(user.isEmpty()){
             throw new ResourceNotFoundException("User with given ID doesn't exist.");
         }
@@ -56,11 +68,19 @@ public class UserService {
         return new UserDTO(updatedUser);
     }
 
-    public void deleteUser (String id){
+    public void deleteUser(String id) {
+        List<Order> ordersByUser = orderRepo.findOrdersByCustomerId(id);
+
+        if (!ordersByUser.isEmpty()) {
+            throw new ConflictException("User with given ID has associated uncompleted orders.");
+        }
+
         Optional<User> user = userRepo.findUserById(id);
-        if(user.isEmpty()){
+        if (user.isEmpty()) {
             throw new ResourceNotFoundException("User with given ID doesn't exist.");
         }
+
         userRepo.delete(user.get());
     }
+
 }
